refactor(components): update Button usage with BaseButton & add ColorHelper to BaseButton
